chore: rename file and move interfaces about multiselect and his input search
