+++ src/types/ui/SearchMultiSelect.ts
@@ -1,0 +1,6 @@
+export interface ISearch {
+  value: string
+  onChange: (event: React.ChangeEvent<HTMLInputElement>) => void
+  placeholder?: string
+  icon?: React.ReactNode
+}
